Delete dead asm!("nop") loop.
Infinite loops without side effects are undefined behavior in LLVM, so
previously, we needed an explicit nop to avoid the loop getting deleted
and our idle task falling through to whatever was below. We now have a
loop with side effects that can't get optimized out, so the loop is now
dead code.